pyproject.toml
[project]
name = "smlmshot"
version = "0.1.0"
description = "Official implementation of the paper 'Optimal transport unlocks end-to-end learning for single-molecule localization'."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "lightning>=2.5.5",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pre-commit>=4.3.0",
    "prodigyopt>=1.1.2",
    "pytest>=8.4.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "sigfig>=1.3.19",
    "tensorboard>=2.20.0",
    "tifffile>=2025.10.4",
    "torch>=2.8.0",
    "torchmetrics>=1.8.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff.lint]
select = ["D", "I", "W", "E", "F"]
ignore = ["D100", "D104", "D105", "D107", "D203", "D213"]

[tool.uv.sources]
torch = [{ index = "pytorch_index" }]

[[tool.uv.index]]
name = "pytorch_index"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
